+++ packages/eslint-plugin-stories/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- [breaking] Remove `no-jest-in-stories`
+- [breaking] Remove `stories-default-export`
+
 ## 2.0.1 (2021-11-29)
 
 - [maintenance] Revert https://github.com/chanzuckerberg/frontend-libs/pull/19 and go back to CommonJS modules (technicalyl this is breaking, but there are no downloads of v3.0.0)